+++ lib/bolt/cli/src/commands/config.rs
@@ -21,6 +21,12 @@ pub enum SubCommand {
 	/// Adds missing regions from supported cloud providers to default_regions.toml.
 	#[clap(hide(true))]
 	GenerateDefaultRegions,
+	ServiceDependencies {
+		#[clap(index = 1)]
+		svc_name: String,
+		#[clap(long, short = 'r')]
+		recursive: bool,
+	},
 }
 
 impl SubCommand {
@@ -34,6 +40,29 @@ impl SubCommand {
 				tasks::config::set_namespace(&namespace).await?;
 			}
 			Self::GenerateDefaultRegions => tasks::config::generate_default_regions().await?,
+			Self::ServiceDependencies {
+				svc_name,
+				recursive,
+			} => {
+				// Build project
+				let ctx = bolt_core::context::ProjectContextData::new(
+					std::env::var("BOLT_NAMESPACE").ok(),
+				)
+				.await;
+
+				// Read deps
+				let deps = if *recursive {
+					ctx.recursive_dependencies(&[&svc_name]).await
+				} else {
+					let svc = ctx.service_with_name(svc_name).await;
+					svc.dependencies().await
+				};
+
+				// Print deps
+				for dep in deps {
+					println!("{}", dep.name());
+				}
+			}
 		}
 
 		Ok(())