+++ lib/bolt/core/src/dep/terraform/pools.rs
@@ -186,18 +186,8 @@ pub async fn build_pools(ctx: &ProjectContext) -> Result<HashMap<String, Pool>> 
 	);
 
 	let mut svc_roles = vec!["docker", "nomad-client", "consul-client"];
-	// Add logging roles
-	if ctx.ns().logging.is_some()
-		&& ctx
-			.read_secret_opt(&["cloudflare", "access", "proxy", "client_id"])
-			.await?
-			.is_some()
-		&& ctx
-			.read_secret_opt(&["cloudflare", "access", "proxy", "client_secret"])
-			.await?
-			.is_some()
-	{
-		svc_roles.extend(["traefik-cloudflare-proxy", "docker-plugin-loki"]);
+	if ctx.ns().logging.is_some() {
+		svc_roles.extend(["traefik", "cloudflare-proxy", "docker-plugin-loki"]);
 	}
 	pools.insert(
 		"svc".into(),
@@ -469,7 +459,7 @@ pub async fn build_pools(ctx: &ProjectContext) -> Result<HashMap<String, Pool>> 
 	pools.insert(
 		"ing-px".into(),
 		PoolBuilder::default()
-			.roles(vec!["traefik", "consul-client"])
+			.roles(vec!["traefik", "ingress-proxy", "consul-client"])
 			.vpc(true)
 			.local_mode(PoolLocalMode::Locally)
 			.tunnels(hashmap! {
@@ -549,7 +539,7 @@ pub async fn build_pools(ctx: &ProjectContext) -> Result<HashMap<String, Pool>> 
 	pools.insert(
 		"ing-job".into(),
 		PoolBuilder::default()
-			.roles(vec!["traefik"])
+			.roles(vec!["traefik", "ingress-proxy"])
 			.vpc(false)
 			.local_mode(PoolLocalMode::Keep)
 			.tunnels(hashmap! {
@@ -659,7 +649,7 @@ pub async fn build_pools(ctx: &ProjectContext) -> Result<HashMap<String, Pool>> 
 				},
 				FirewallRule {
 					label: "nomad-dynamic-udp".into(),
-					ports: "20000-25999".into(),
+					ports: "20000-31999".into(),
 					protocol: "udp".into(),
 					inbound_ipv4_cidr: vec!["0.0.0.0/0".into()],
 					inbound_ipv6_cidr: vec!["::/0".into()],