+++ lib/bolt/core/src/tasks/infra/mod.rs
@@ -16,7 +16,13 @@ pub struct ExecutePlanOpts {
 }
 
 #[derive(Debug, Clone)]
-pub enum PlanStep {
+pub struct PlanStep {
+	pub name_id: &'static str,
+	pub kind: PlanStepKind,
+}
+
+#[derive(Debug, Clone)]
+pub enum PlanStepKind {
 	Terraform {
 		plan_id: String,
 
@@ -38,10 +44,10 @@ pub enum PlanStep {
 	Up,
 }
 
-impl PlanStep {
+impl PlanStepKind {
 	async fn execute(&self, ctx: ProjectContext, opts: &ExecutePlanOpts) -> Result<()> {
 		match self {
-			PlanStep::Terraform { plan_id, .. } => {
+			PlanStepKind::Terraform { plan_id, .. } => {
 				let varfile_path = ctx.gen_tf_env_path();
 
 				let mut cmd = terraform::cli::build_command(&ctx, plan_id).await;
@@ -55,7 +61,7 @@ impl PlanStep {
 
 				terraform::output::clear_cache(&ctx, &plan_id).await;
 			}
-			PlanStep::Salt {
+			PlanStepKind::Salt {
 				filter,
 				sls,
 				config_opts,
@@ -70,10 +76,10 @@ impl PlanStep {
 					salt::cli::apply_all(&ctx, &apply_opts, config_opts).await?;
 				}
 			}
-			PlanStep::Migrate => {
+			PlanStepKind::Migrate => {
 				tasks::migrate::up_all(&ctx).await?;
 			}
-			PlanStep::Up => {
+			PlanStepKind::Up => {
 				tasks::up::up_all(
 					&ctx,
 					tasks::up::UpOpts {
@@ -90,7 +96,7 @@ impl PlanStep {
 
 	async fn destroy(&self, ctx: ProjectContext, opts: &ExecutePlanOpts) -> Result<()> {
 		match self {
-			PlanStep::Terraform {
+			PlanStepKind::Terraform {
 				plan_id,
 				needs_destroy,
 			} => {
@@ -111,7 +117,7 @@ impl PlanStep {
 
 				terraform::output::clear_cache(&ctx, &plan_id).await;
 			}
-			PlanStep::Salt { .. } | PlanStep::Migrate | PlanStep::Up => {
+			PlanStepKind::Salt { .. } | PlanStepKind::Migrate | PlanStepKind::Up => {
 				// Do nothing
 			}
 		}
@@ -120,70 +126,97 @@ impl PlanStep {
 	}
 }
 
-pub fn build_plan(ctx: &ProjectContext) -> Result<Vec<PlanStep>> {
+pub fn build_plan(ctx: &ProjectContext, start_at: Option<String>) -> Result<Vec<PlanStep>> {
 	let mut plan = Vec::new();
 
 	// TLS
-	plan.push(PlanStep::Terraform {
-		plan_id: "tls".into(),
-		needs_destroy: true,
+	plan.push(PlanStep {
+		name_id: "tf-tls",
+		kind: PlanStepKind::Terraform {
+			plan_id: "tls".into(),
+			needs_destroy: true,
+		},
 	});
 
 	// Nebula
-	plan.push(PlanStep::Terraform {
-		plan_id: "nebula".into(),
-		needs_destroy: false,
+	plan.push(PlanStep {
+		name_id: "tf-nebula",
+		kind: PlanStepKind::Terraform {
+			plan_id: "nebula".into(),
+			needs_destroy: false,
+		},
 	});
 
 	// Master
 	match ctx.ns().cluster.kind {
 		ClusterKind::SingleNode { .. } => {
-			plan.push(PlanStep::Terraform {
-				plan_id: "master_local".into(),
-				needs_destroy: false,
+			plan.push(PlanStep {
+				name_id: "tf-master-local",
+				kind: PlanStepKind::Terraform {
+					plan_id: "master_local".into(),
+					needs_destroy: false,
+				},
 			});
 		}
 		ClusterKind::Distributed { .. } => {
-			plan.push(PlanStep::Terraform {
-				plan_id: "master_cluster".into(),
-				needs_destroy: true,
+			plan.push(PlanStep {
+				name_id: "tf-master-cluster",
+				kind: PlanStepKind::Terraform {
+					plan_id: "master_cluster".into(),
+					needs_destroy: true,
+				},
 			});
 		}
 	}
 
 	// Pools
-	plan.push(PlanStep::Terraform {
-		plan_id: "pools".into(),
-		needs_destroy: true,
+	plan.push(PlanStep {
+		name_id: "tf-pools",
+		kind: PlanStepKind::Terraform {
+			plan_id: "pools".into(),
+			needs_destroy: true,
+		},
 	});
 
 	// DNS
-	plan.push(PlanStep::Terraform {
-		plan_id: "dns".into(),
-		needs_destroy: true,
+	plan.push(PlanStep {
+		name_id: "tf-dns",
+		kind: PlanStepKind::Terraform {
+			plan_id: "dns".into(),
+			needs_destroy: true,
+		},
 	});
 
 	// Cloudflare
-	plan.push(PlanStep::Terraform {
-		plan_id: "cloudflare_workers".into(),
-		needs_destroy: true,
+	plan.push(PlanStep {
+		name_id: "tf-cf-workers",
+		kind: PlanStepKind::Terraform {
+			plan_id: "cloudflare_workers".into(),
+			needs_destroy: true,
+		},
 	});
 
 	if let config::ns::DnsProvider::Cloudflare {
 		access: Some(_), ..
 	} = ctx.ns().dns.provider
 	{
-		plan.push(PlanStep::Terraform {
-			plan_id: "cloudflare_tunnels".into(),
-			needs_destroy: true,
+		plan.push(PlanStep {
+			name_id: "tf-cf-tunnels",
+			kind: PlanStepKind::Terraform {
+				plan_id: "cloudflare_tunnels".into(),
+				needs_destroy: true,
+			},
 		});
 	}
 
 	// Grafana
 	if ctx.ns().grafana.is_some() {
-		plan.push(PlanStep::Terraform {
-			plan_id: "grafana".into(),
-			needs_destroy: true,
+		plan.push(PlanStep {
+			name_id: "tf-grafana",
+			kind: PlanStepKind::Terraform {
+				plan_id: "grafana".into(),
+				needs_destroy: true,
+			},
 		});
 	}
 
@@ -191,55 +224,89 @@ pub fn build_plan(ctx: &ProjectContext) -> Result<Vec<PlanStep>> {
 	let s3_providers = &ctx.ns().s3.providers;
 	if s3_providers.minio.is_some() {
 		// Install Minio for s3_minio Terraform plan
-		plan.push(PlanStep::Salt {
-			filter: Some("G@roles:minio".into()),
-			sls: None,
-			config_opts: salt::config::BuildOpts { skip_s3: true },
+		plan.push(PlanStep {
+			name_id: "salt-minio",
+			kind: PlanStepKind::Salt {
+				filter: Some("G@roles:minio".into()),
+				sls: None,
+				config_opts: salt::config::BuildOpts { skip_s3: true },
+			},
 		});
 
-		plan.push(PlanStep::Terraform {
-			plan_id: "s3_minio".into(),
-			needs_destroy: false,
+		plan.push(PlanStep {
+			name_id: "tf-s3-minio",
+			kind: PlanStepKind::Terraform {
+				plan_id: "s3_minio".into(),
+				needs_destroy: false,
+			},
 		});
 	}
 	if s3_providers.backblaze.is_some() {
-		plan.push(PlanStep::Terraform {
-			plan_id: "s3_backblaze".into(),
-			needs_destroy: true,
+		plan.push(PlanStep {
+			name_id: "tf-s3-backblaze",
+			kind: PlanStepKind::Terraform {
+				plan_id: "s3_backblaze".into(),
+				needs_destroy: true,
+			},
 		});
 	}
 	if s3_providers.aws.is_some() {
-		plan.push(PlanStep::Terraform {
-			plan_id: "s3_aws".into(),
-			needs_destroy: true,
+		plan.push(PlanStep {
+			name_id: "tf-s3-aws",
+			kind: PlanStepKind::Terraform {
+				plan_id: "s3_aws".into(),
+				needs_destroy: true,
+			},
 		});
 	}
 
 	// Apply the rest of the Salt configs
-	plan.push(PlanStep::Salt {
-		filter: None,
-		sls: None,
-		config_opts: Default::default(),
+	plan.push(PlanStep {
+		name_id: "salt",
+		kind: PlanStepKind::Salt {
+			filter: None,
+			sls: None,
+			config_opts: Default::default(),
+		},
 	});
 
-	plan.push(PlanStep::Terraform {
-		plan_id: "nomad".into(),
-		needs_destroy: false,
+	plan.push(PlanStep {
+		name_id: "tf-nomad",
+		kind: PlanStepKind::Terraform {
+			plan_id: "nomad".into(),
+			needs_destroy: false,
+		},
 	});
 
-	plan.push(PlanStep::Migrate);
+	plan.push(PlanStep {
+		name_id: "migrate",
+		kind: PlanStepKind::Migrate,
+	});
 
-	plan.push(PlanStep::Up);
+	plan.push(PlanStep {
+		name_id: "up",
+		kind: PlanStepKind::Up,
+	});
+
+	// Start at the specified step
+	if let Some(start_at) = start_at {
+		let idx = plan
+			.iter()
+			.position(|x| x.name_id == start_at)
+			.ok_or_else(|| anyhow!("invalid start_at value: {}", start_at))?;
+
+		plan = plan[idx..].to_vec();
+	}
 
 	Ok(plan)
 }
 
 /// List all of the Terraform plans in use for the generated plan.
 pub fn all_terraform_plans(ctx: &ProjectContext) -> Result<Vec<String>> {
-	let plan_ids = build_plan(ctx)?
+	let plan_ids = build_plan(ctx, None)?
 		.into_iter()
 		.flat_map(|x| {
-			if let PlanStep::Terraform { plan_id, .. } = x {
+			if let PlanStepKind::Terraform { plan_id, .. } = x.kind {
 				Some(plan_id)
 			} else {
 				None
@@ -260,8 +327,11 @@ pub async fn execute_plan(
 	for (i, step) in plan.iter().enumerate() {
 		eprintln!();
 		eprintln!();
-		rivet_term::status::info("Executing", format!("({}/{}) {step:?}", i + 1, plan.len()));
-		step.execute(ctx.clone(), &opts).await?;
+		rivet_term::status::info(
+			"Executing",
+			format!("({}/{}) {}", i + 1, plan.len(), step.name_id),
+		);
+		step.kind.execute(ctx.clone(), &opts).await?;
 	}
 
 	Ok(())
@@ -277,8 +347,11 @@ pub async fn destroy_plan(
 	for (i, step) in plan.iter().enumerate().rev() {
 		eprintln!();
 		eprintln!();
-		rivet_term::status::info("Destroying", format!("({}/{}) {step:?}", i + 1, plan.len()));
-		step.destroy(ctx.clone(), &opts).await?;
+		rivet_term::status::info(
+			"Destroying",
+			format!("({}/{}) {}", i + 1, plan.len(), step.name_id),
+		);
+		step.kind.destroy(ctx.clone(), &opts).await?;
 	}
 
 	Ok(())