+++ svc/pkg/mm-config/ops/version-publish/src/lib.rs
@@ -42,7 +42,7 @@ async fn handle(
 	.bind(version_id)
 	.bind(captcha_buf)
 	.bind(util_mm::version_migrations::all())
-	.execute(&mut tx)
+	.execute(&mut *tx)
 	.await?;
 
 	// Save lobby groups
@@ -128,6 +128,7 @@ async fn handle(
 		.bind(&find_config_buf)
 		.bind(&join_config_buf)
 		.execute(&mut tx)
+		.execute(&mut *tx)
 		.await?;
 
 		for region in &lobby_group.regions {
@@ -143,7 +144,7 @@ async fn handle(
 			.bind(lobby_group_id)
 			.bind(region_id)
 			.bind(&region.tier_name_id)
-			.execute(&mut tx)
+			.execute(&mut *tx)
 			.await?;
 
 			if let Some(idle_lobbies) = &region.idle_lobbies {
@@ -159,7 +160,7 @@ async fn handle(
 				.bind(region_id)
 				.bind(idle_lobbies.min_idle_lobbies as i64)
 				.bind(idle_lobbies.max_idle_lobbies as i64)
-				.execute(&mut tx)
+				.execute(&mut *tx)
 				.await?;
 			}
 		}