+++ svc/pkg/mm/worker/src/workers/lobby_find/mod.rs
@@ -592,7 +592,7 @@ async fn insert_to_crdb(
 		.bind(lobby_group_config.lobby_group.max_players_normal as i64)
 		.bind(lobby_group_config.lobby_group.max_players_direct as i64)
 		.bind(lobby_group_config.lobby_group.max_players_party as i64)
-		.execute(&mut *tx)
+		.execute(&mut **tx)
 		.await?;
 	}
 
@@ -616,7 +616,7 @@ async fn insert_to_crdb(
 	.bind(lobby_id)
 	.bind(auto_create_lobby)
 	.bind(util_mm::FindQueryStatus::Pending as i64)
-	.execute(&mut *tx)
+	.execute(&mut **tx)
 	.await?;
 
 	// Insert players
@@ -647,7 +647,7 @@ async fn insert_to_crdb(
 		)
 		.bind(now_ts)
 		.bind(ray_id)
-		.execute(&mut *tx)
+		.execute(&mut **tx)
 		.await?;
 	}