+++ CHANGELOG.md
@@ -9,12 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   **Cloud** Support multipart uploads for builds
 -   **Infra** Support configuring multiple S3 providers
 -   **Infra** Support multipart uploads
 -   **Infra** Replace Promtail-based log shipping with native Loki Docker driver
--   **Infra** Add local Traefik Cloudflare proxy daemon for connecting to Cloudflare Access services
+-   **Infra** Local Traefik Cloudflare proxy daemon for connecting to Cloudflare Access services
+-   **Infra** Upload service builds to default S3 provider instead of hardcoded bucket
 -   **Bolt** Support for connecting to Redis databases with `bolt redis sh`
--   **Bolt** Add confirmation before running any command in the production namespace
+-   **Bolt** Confirmation before running any command in the production namespace
+-   **Bolt** `--start-at` flag for all infra commands
 
 ### Changed
 
@@ -25,12 +28,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   **Infra** Update Consul to 1.16.0
 -   **Infra** Update Imagor to 1.4.7
 -   **Infra** Update NATS server to 2.9.20
--   **Infra** Update Node Exporter server to 1.6.1
--   **Infra** Update Nomad to 1.6.1
+-   **Infra** Update Node Exporter server to 1.6.0
+-   **Infra** Update Nomad to 1.6.0
 -   **Infra** Update Prometheus server to 2.46.0
 -   **Infra** Update Redis Exporter to 1.52.0
 -   **Infra** Update Redis to 7.0.12
 -   **Infra** Update Treafik to 2.10.4
+-   **Bolt** PostHog events are now captured in a background task
+-   **Bolt** Auto-install rsync on Salt Master
+-   **Bolt** Recursively add dependencies from overridden services when using additional roots
+-   **KV** Significantly rate limit of all endpoints
 
 ### Security
 
@@ -40,4 +47,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   **Portal** Skip captcha if no Turnstile key provided
 -   **Infra** Resolve [RUSTSEC-2023-0044](https://rustsec.org/advisories/RUSTSEC-2023-0044)
+-   **Infra** Missing dpenedency on mounting volumn before setting permissions of /var/* for Cockroach, ClickHouse, Prometheus, and Traffic Server
+-   **Chrip** Empty message parameters now have placeholder so NATS doesn't throw an error
+-   **Chrip** Messages with no parameters no longer have a trailing dot
 -   **Bolt** Correctly resolve project root when building services natively
+-   **Bolt** Correctly determine executable path for `ExecServiceDriver::UploadedBinaryArtifact` with different Cargo names
+